+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-01-16-2
+
+Changes:
+
+- `HttpRequest` / `HttpResponse`
+
 ## v2023-01-16-1
 
 Changes: